lib: change message length to uint16 for OFDM and FSK compatibility
Change CSP to match type change
Add API to manage retro compatibility with old CSP version

+++ connect/ember-types.h
@@ -209,7 +209,7 @@ enum
 /**
  * @brief Message length in bytes.
  */
-typedef uint8_t EmberMessageLength;
+typedef uint16_t EmberMessageLength;
 
 /**
  * @brief Message options.
